requirements.txt
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-nspkg==3.0.2
azure-storage-blob==12.20.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
Flask==3.0.3
google-auth==2.32.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pymongo==4.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
